requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Keras == 2.4.3
Pillow == 8.0.1
matplotlib == 3.3.2
numpy == 1.20.1
opencv-contrib-python == 4.5.1.48
opencv-contrib-python-headless == 4.5.1.48
opencv-openvino-contrib-python == 4.1.1.26
opencv-python == 4.5.1.48
opencv-python-aarch64 == 3.3.0.1
opencv-python-headless == 4.5.1.48
opencv-python-inference-engine == 4.5.0.1
opencvz-contrib-python == 4.1.1.26
sklearn == 0.0
tensorflow == 2.4.1
tensorflow-hub == 0.11.0
tensorflow_datasets == 4.2.0
tf-hub-nightly == 0.12.0.dev202102250005
progress == 1.5
